+++ godotopenxrkhronos/src/main/java/org/godotengine/openxr/vendors/khronos/GodotOpenXRKhronos.kt
@@ -29,6 +29,7 @@
 
 package org.godotengine.openxr.vendors.khronos
 
+import android.util.Log
 import org.godotengine.godot.Godot
 import org.godotengine.godot.plugin.GodotPlugin
 
@@ -49,6 +50,21 @@ import org.godotengine.godot.plugin.GodotPlugin
  *
  */
 class GodotOpenXRKhronos(godot: Godot?) : GodotPlugin(godot) {
+    companion object {
+        private val TAG = GodotOpenXRKhronos::class.java.simpleName
+
+        init {
+            try {
+                Log.v(TAG, "Loading godotopenxrvendors library")
+                System.loadLibrary("godotopenxrvendors")
+            } catch (e: UnsatisfiedLinkError) {
+                Log.e(TAG, "Unable to load godotopenxrvendors shared library")
+            }
+        }
+    }
+
+    override fun getPluginGDExtensionLibrariesPaths() = setOf("res://addons/godotopenxrvendors/plugin.gdextension")
+
     override fun getPluginName(): String {
         return "GodotOpenXRKhronos"
     }